move conversation sidebar inside full screen folder
