+++ x-pack/platform/plugins/shared/onechat/public/application/components/conversations/full_screen/conversation_sidebar/no_conversations_prompt.tsx
